move mtgogetter function to submodule
