+++ mtgoupdater/src/lib.rs
@@ -2,6 +2,8 @@
 use std::process::Command;
 use std::sync::OnceLock;
 
+mod mtgogetter_api;
+
 static MTGOGETTER_BIN: OnceLock<String> = OnceLock::new();
 static MTGOPARSER_BIN: OnceLock<String> = OnceLock::new();
 
@@ -13,11 +15,11 @@ pub fn set_mtgoparser_bin(bin_path: &str) -> Result<(), String> {
     MTGOPARSER_BIN.set(bin_path.to_string())
 }
 
-fn mtgogetter_bin() -> &'static str {
+pub(crate) fn mtgogetter_bin() -> &'static str {
     MTGOGETTER_BIN.get().expect("MTGOGETTER_BIN not set")
 }
 
-fn mtgoparser_bin() -> &'static str {
+pub(crate) fn mtgoparser_bin() -> &'static str {
     MTGOPARSER_BIN.get().expect("MTGOPARSER_BIN not set")
 }
 
@@ -48,44 +50,6 @@ pub mod internal_only {
     } else {
         "../mtgoparser/build/src/mtgo_preprocessor/Release/mtgo_preprocesser"
     };
-}
-
-pub fn download_goatbots_price_history() -> Result<std::process::Output, Box<dyn std::error::Error>>
-{
-    let go_exec_out = Command::new(mtgogetter_bin())
-        .arg("download")
-        .arg("goatbots-price-history")
-        .output()?;
-
-    Ok(go_exec_out)
-}
-
-pub fn download_goatbots_card_definitions(
-) -> Result<std::process::Output, Box<dyn std::error::Error>> {
-    let go_exec_out = Command::new(mtgogetter_bin())
-        .arg("download")
-        .arg("goatbots-card-definitions")
-        .output()?;
-
-    Ok(go_exec_out)
-}
-
-pub fn download_custom_url(
-    url: &str,
-    decompress: bool,
-    save_as: Option<&str>,
-) -> Result<std::process::Output, Box<dyn std::error::Error>> {
-    let mut custom_args = vec!["download", "custom", "url-raw", url];
-    if decompress {
-        custom_args.push("--decompress");
-    }
-    if let Some(save_as) = save_as {
-        custom_args.push("--save-as");
-        custom_args.push(save_as);
-    }
-    let go_exec_out = Command::new(mtgogetter_bin()).args(custom_args).output()?;
-
-    Ok(go_exec_out)
 }
 
 pub fn run_mtgo_preprocessor_example() -> Result<std::process::Output, Box<dyn std::error::Error>> {
